Update on "Don't use NonVariableTypeMode in custom ops"
Potentially fixes https://github.com/pytorch/pytorch/issues/37306

Differential Revision: [D21261946](https://our.internmc.facebook.com/intern/diff/D21261946/)

[ghstack-poisoned]
